add infrastructure for intermediate readers
This is a draft of a infrastructure for Reader->Reader intermediate
stages. The input is a flat mutation reader (either devirtualised or a
flat_mutation_reader) and the output is an implementation-dependent
type. Reader->Reader stages can be composed and the final result can be
type-erased producing a flat_mutation_reader.