tests/perf_fast_forward: create data directory in /tmp by default
perf_fast_forward populates perf_fast_forward_output with some data and
then runs performance tests that read it. That makes the disk a
significant factor in the final result and makes the results less
repeatable. The simples way to remove that varialbe is to use ramdisk
instead of a real disk.