converting_mutation_partition_applier: do not emit empty collections
This patch changes the behaviour of the schema upgrade code so that if
all cells and the tombstons of a collection a removed during the upgrade
the collection is not emitted (as opposed to emitting an empty one).
Both behaviours are valid, but the new one makes it more consistent with
how atomic cells are upgraded and how schema upgrades work for sstable
readers.