Revert "mutation_source_test: add option to skip intra-partition fast-forwarding tests"
This reverts commit b36733971b1298c36d7a418966ec5704f27f2bd3. That commit made
run_mutation_reader_tests() support  mutation_sources that do not implement
streamed_mutation::forwarding::yes. This is wrong since mutation_sources
are not allowed to ignore or otherwise not support that mode. Moreover,
there is absolutely no reason for them to do so since there is a
make_forwardable() adapter that can make any mutation_reader a
forwardable one (at the cost of performance, but that's not always
important).