encoding_stats: add column information
The sstables mc format can save space if it is guaranteed that no
mutation contains certain columns. This information is kept in two
bitsets in encoding_stats (one for static columns, one for regular
ones). The default-constructed (i.e. empty) bitset indicate that all
columns are present which is a safe, albeit suboptimal, default.