sstables/mc/writer: write only columns present in encoding_stats
This patch makes the mc sstable writer use encoding statistics to
completely skip columns that are guaranteed to be not present in any of
the mutations that are about to be written.

Some of sstable tests compare the output with golden sstables. Since we
now skip unused columns those need to be updated.