commitlog::segment: use vectored IO when writing fragmented buffers
At the moment, when it is time to write a fragmented buffer, cycle()
issues a separate dma_write() call for each fragment. This is correct,
but not optimal. This patch changes the code to use vectored IO instead
so that the whole buffer can be written in a single dma_write() call.